Shapes: Moved Blender convertor to a BlenderMesh plugin
