+++ include/mitsuba/render/mesh.h
@@ -35,13 +35,6 @@ public:
          ScalarSize face_count, const Properties &props = Properties(),
          bool has_vertex_normals = false, bool has_vertex_texcoords = false);
 
-    /// Create a new mesh from a blender mesh
-    Mesh(const std::string &name,
-        uintptr_t loop_tri_count, uintptr_t loop_tri_ptr,
-        uintptr_t loop_ptr, uintptr_t vertex_count, uintptr_t vertex_ptr,
-        uintptr_t poly_ptr, uintptr_t uv_ptr, uintptr_t col_ptr,
-        short mat_nr, const ScalarTransform4f &to_world,
-        const Properties &props = Properties());
     // =========================================================================
     //! @{ \name Accessors (vertices, faces, normals, etc)
     // =========================================================================
