+++ src/shapes/CMakeLists.txt
@@ -2,6 +2,7 @@ set(MTS_PLUGIN_PREFIX "shapes")
 
 add_plugin(obj         obj.cpp)
 add_plugin(ply         ply.cpp)
+add_plugin(blender     blender.cpp)
 add_plugin(serialized  serialized.cpp)
 
 add_plugin(cylinder    cylinder.cpp)
